perf(fdb): use ByteArrayUtil encodeInt/decodeInt for segment id pointers in FdbVectorStore and maxSeg read in VectorIndex
- Replace Tuple pack/unpack for currentSegment and maxSegment pointers.
- Keep tuple for structured segment subkeys; only pointers switch to raw int encoding.

+++ src/main/java/io/github/panghy/vectorsearch/api/VectorIndex.java
@@ -289,7 +289,7 @@ public class VectorIndex implements AutoCloseable {
     byte[] maxK = dirs.maxSegmentKey();
     return db.readAsync(tr -> tr.get(maxK).thenCompose(maxB -> {
       if (maxB == null) return completedFuture(List.of());
-      int maxSeg = (int) decodeInt(maxB);
+      int maxSeg = Math.toIntExact(decodeInt(maxB));
       // Batch get meta keys for [0..maxSeg]
       List<byte[]> keys = new ArrayList<>(maxSeg + 1);
       for (int i = 0; i <= maxSeg; i++)
